Create separate package for network responses.
