+++ app/src/main/java/ru/sandbox/androidacademyapp/data/network/responses/GenreResponse.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp.data.network
+package ru.sandbox.androidacademyapp.data.network.responses
 
 import kotlinx.serialization.Serializable
